--- src/saguaro-3d/external-alignment-provider.ts
@@ -1,244 +1,0 @@
-import {
-    AlignmentResponse,
-} from '@rcsb/rcsb-api-tools/build/RcsbGraphQL/Types/Borrego/GqlTypes';
-import {
-    AlignmentCollectorInterface
-} from '@rcsb/rcsb-saguaro-app/build/dist/RcsbCollectTools/AlignmentCollector/AlignmentCollectorInterface';
-import { TagDelimiter } from '@rcsb/rcsb-api-tools/build/RcsbUtils/TagDelimiter';
-
-
-import { AlignmentMapType, AlignmentReference } from './alignment-reference';
-import {
-    LoadParamsProviderInterface,
-    RigidTransformType,
-    TransformMatrixType
-} from '@rcsb/rcsb-saguaro-3d/lib//RcsbFvStructure/StructureUtils/StructureLoaderInterface';
-
-import {
-    StructureAlignmentResponse,
-    StructureEntry,
-    StructureURL
-} from '../auto/alignment/alignment-response';
-import {
-    LoadMethod,
-    LoadMolstarInterface,
-    LoadMolstarReturnType
-} from '@rcsb/rcsb-saguaro-3d/lib//RcsbFvStructure/StructureViewers/MolstarViewer/MolstarActionManager';
-import {
-    AlignmentTrajectoryParamsType, AlignmentTrajectoryPresetProvider
-} from './molstar-trajectory/alignment-trajectory-preset-provider';
-import { ColorLists } from '../utils/color';
-import { DefaultQueryRuntimeTable, QuerySymbolRuntime } from 'molstar/lib/mol-script/runtime/query/compiler';
-import { CustomPropSymbol } from 'molstar/lib/mol-script/language/symbol';
-import { Type } from 'molstar/lib/mol-script/language/type';
-import { StructureProperties } from 'molstar/lib/mol-model/structure/structure/properties';
-import { CustomPropertyDescriptor } from 'molstar/lib/mol-model/custom-property';
-
-export class RcsbStructuralAlignmentProvider implements AlignmentCollectorInterface {
-
-    private alignmentResponse: AlignmentResponse | undefined = undefined;
-    private readonly alignment: StructureAlignmentResponse;
-    private readonly alignmentReference: AlignmentReference;
-    constructor(alignment: StructureAlignmentResponse, alignmentReference: AlignmentReference) {
-        this.alignment = alignment;
-        this.alignmentReference = alignmentReference;
-    }
-
-    async collect(): Promise<AlignmentResponse> {
-        return await this.data();
-    }
-    async getTargets(): Promise<string[]> {
-        return (await this.data()).target_alignment?.map(ta=>ta?.target_id ?? 'NA') ?? [];
-    }
-    async getAlignmentLength(): Promise<number> {
-        return Math.max(...(await this.data()).target_alignment?.map(ta=>ta?.aligned_regions?.[ta?.aligned_regions?.length - 1]?.query_end ?? -1) ?? []);
-    }
-    async getAlignment(): Promise<AlignmentResponse> {
-        return await this.data();
-    }
-    private async data(): Promise<AlignmentResponse> {
-        if (!this.alignmentResponse)
-            this.alignmentResponse = await alignmentTransform(this.alignment, this.alignmentReference);
-        return this.alignmentResponse;
-    }
-
-}
-
-class RcsbStructuralTransformProvider {
-
-    private readonly alignment: StructureAlignmentResponse;
-    constructor(alignment: StructureAlignmentResponse) {
-        this.alignment = alignment;
-    }
-
-    get(alignmentIndex: number, pairIndex: number): RigidTransformType[] {
-
-        const res = this.alignment.results?.[alignmentIndex];
-        if (res?.structure_alignment.length === 1) {
-            return [{
-                transform: res.structure_alignment[0].transformations[pairIndex] as TransformMatrixType
-            }];
-        } else if (res?.structure_alignment.length && res?.structure_alignment.length > 1) {
-            return res.structure_alignment.map(sa=>({
-                transform: sa.transformations[pairIndex] as TransformMatrixType,
-                regions: sa.regions?.[pairIndex].map(r=>[r.beg_seq_id, r.beg_seq_id + r.length - 1])
-            }));
-        } else {
-            return [{
-                transform: [1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1]
-            }];
-        }
-    }
-
-}
-
-export class AlignmentColoringConfig {
-
-    private readonly colorConfig: {closeResidues: Map<string, Set<number>>; colors: Map<string, number>;};
-    readonly idMap: Map<string, string> = new Map<string, string>();
-    private readonly uniqueChainMap: Map<string, {asymId: string; operatorName: string;}> = new Map<string, {asymId: string; operatorName: string;}>();
-    constructor(colorConfig: {closeResidues: Map<string, Set<number>>; colors: Map<string, number>;}) {
-        this.colorConfig = colorConfig;
-    }
-
-    public getCloseResidues(modelId: string): Set<number> {
-        return this.colorConfig.closeResidues.get(this.idMap.get(modelId) ?? 'none') ?? new Set<number>();
-    }
-
-    public getModelColor(modelId: string): number {
-        return this.colorConfig.colors.get(this.idMap.get(modelId) ?? 'none') ?? 0x777777;
-    }
-
-    public setAlignmentIdToModel(modelId: string, alignmentId: string): void {
-        this.idMap.set(modelId, alignmentId);
-    }
-
-    public setUniqueChain(modelId: string, asymId: string, operatorName: string): void {
-        this.uniqueChainMap.set(modelId, { asymId, operatorName });
-    }
-
-    public getUniqueChain(modelId: string): {asymId: string; operatorName: string;} | undefined {
-        return this.uniqueChainMap.get(modelId);
-    }
-
-}
-
-export class RcsbLoadParamsProvider implements LoadParamsProviderInterface<{entryId: string; instanceId: string;}, LoadMolstarInterface<AlignmentTrajectoryParamsType, LoadMolstarReturnType>> {
-
-    private readonly alignment: StructureAlignmentResponse;
-    private readonly transformProvider: RcsbStructuralTransformProvider;
-    private readonly alignmentReference: AlignmentReference;
-
-    constructor(alignment: StructureAlignmentResponse, alignmentReference: AlignmentReference) {
-        this.alignment = alignment;
-        this.transformProvider = new RcsbStructuralTransformProvider(alignment);
-        this.alignmentReference = alignmentReference;
-    }
-
-    get(pdb: {entryId: string; instanceId: string;}): LoadMolstarInterface<AlignmentTrajectoryParamsType, LoadMolstarReturnType> {
-        if (!this.alignment.results)
-            throw new Error('Alignments results not found');
-        const { alignmentIndex, pairIndex, entryId } = this.alignmentReference.getAlignmentEntry(`${pdb.entryId}${TagDelimiter.instance}${pdb.instanceId}`);
-        const alignmentId = `${pdb.entryId}${TagDelimiter.instance}${pdb.instanceId}`;
-        const res = this.alignment.results[alignmentIndex];
-        const structure = res.structures[pairIndex] as StructureEntry & StructureURL;
-        const transform = this.transformProvider.get(alignmentIndex, pairIndex);
-        const reprProvider = AlignmentTrajectoryPresetProvider;
-        const loadMethod = 'url' in structure && structure.url ? LoadMethod.loadStructureFromUrl : LoadMethod.loadPdbId;
-        const url: string|undefined = 'url' in structure && structure.url ? structure.url : undefined;
-        return {
-            loadMethod,
-            loadParams: {
-                url,
-                entryId,
-                format: structure.format,
-                isBinary: structure.is_binary,
-                id: alignmentId,
-                reprProvider: reprProvider,
-                params: {
-                    pdb,
-                    alignmentId,
-                    modelIndex: 0,
-                    transform: transform,
-                    targetAlignment: undefined
-                }
-            }
-        };
-    }
-}
-
-async function alignmentTransform(alignment: StructureAlignmentResponse, alignmentRef: AlignmentReference): Promise<AlignmentResponse> {
-    if (!alignment.results)
-        return {};
-    const out: AlignmentResponse = alignmentRef.buildAlignments();
-    const seqs = await alignmentRef.getSequences();
-    out.target_alignment?.forEach(ta=>{
-        const seq = seqs.find(s=>s.rcsbId === ta?.target_id)?.sequence;
-        if (seq && ta)
-            ta.target_sequence = seq;
-    });
-    return out;
-}
-
-export function alignmentCloseResidues(results: AlignmentMapType[]): Map<string, Set<number>> {
-    const out: Map<string, Set<number>> = new Map<string, Set<number>>();
-    results.slice(1).forEach(alignment=>{
-        const res = alignment.alignment;
-        const instanceId = alignment.alignmentId;
-        if (!out.has(instanceId))
-            out.set(instanceId, new Set<number>());
-        res.structure_alignment.map(sa => sa.regions?.[1]).flat().forEach(reg=>{
-            if (!reg)
-                return;
-            for (let n = 0; n < reg.length; n++) {
-                out.get(instanceId)?.add(reg.beg_seq_id + n);
-            }
-        });
-        const refId = results[0].alignmentId;
-        if (!out.has(refId))
-            out.set(refId, new Set<number>());
-        res.structure_alignment.map(sa => sa.regions?.[0]).flat().forEach(reg=>{
-            if (!reg)
-                return;
-            for (let n = 0; n < reg.length; n++) {
-                out.get(refId)?.add(reg.beg_seq_id + n);
-            }
-        });
-    });
-    return out;
-}
-
-export function entryColors(results: AlignmentMapType[]): Map<string, number> {
-    const out: Map<string, number> = new Map<string, number>();
-    results.forEach((res, n)=>{
-        const instanceId = res.alignmentId;
-        out.set(instanceId, ColorLists['set-1'][n]);
-    });
-    return out;
-}
-
-export const ColorConfigDescriptor = CustomPropertyDescriptor({
-    name: 'alignment.color-config',
-    symbols: {
-        closeResidue: QuerySymbolRuntime.Dynamic(CustomPropSymbol('rcsb', 'alignment.color-config.close-residue', Type.Bool),
-            ctx => {
-                const { structure } = ctx.element;
-                const colorConfig = structure.inheritedPropertyData.colorConfig as AlignmentColoringConfig | undefined;
-                if (!colorConfig) return false;
-
-                const closeResidues = colorConfig.getCloseResidues(structure.model.id);
-                if (!closeResidues) return false;
-
-                const uniqueChain = colorConfig.getUniqueChain(structure.model.id);
-                if (!uniqueChain) return false;
-
-                const asymId = StructureProperties.chain.label_asym_id(ctx.element);
-                const seqId = StructureProperties.residue.label_seq_id(ctx.element);
-                const operatorName = StructureProperties.unit.operator_name(ctx.element);
-                return uniqueChain.asymId === asymId && uniqueChain.operatorName === operatorName && closeResidues.has(seqId);
-            }
-        )
-    }
-});
-
-DefaultQueryRuntimeTable.addCustomProp(ColorConfigDescriptor);
